requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via aiogram
aiogram==3.22.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via aiohttp
attrs==25.4.0
    # via aiohttp
build==1.3.0
    # via pip-tools
certifi==2025.11.12
    # via aiogram
click==8.3.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via build
pip-tools==7.5.2
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.10
    # via aiogram
pydantic-core==2.33.2
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.2.1
    # via -r requirements.in
tomli==2.3.0
    # via
    #   build
    #   pip-tools
typing-extensions==4.15.0
    # via
    #   aiogram
    #   aiosignal
    #   multidict
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
wheel==0.45.1
    # via pip-tools
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
